backend/requirements.txt
# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
aiosqlite==0.20.0

# PDF Processing
pdfplumber==0.11.4
PyMuPDF==1.24.14
pdf2image==1.17.0

# HTTP Client
httpx==0.28.1
aiohttp==3.11.10

# Utilities
pydantic==2.10.0
pydantic-settings==2.6.1
email-validator==2.2.0
python-dotenv==1.0.1

# Payment (Toss Payments)
python-tosspayments


# Logging
loguru==0.7.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
